Fixed warning on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@27057 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/General/datainpspec.cc
@@ -396,7 +396,7 @@ double StringListInpSpec::getdValue( int idx ) const
 
 
 float StringListInpSpec::getfValue( int idx ) const
-{ return cur_; }
+{ return (float) cur_; }
 
 
 void StringListInpSpec::setValue( int i, int idx )
